pyproject.toml
[project]
name = "auto-offer"
version = "0.1.0"
description = "半自动简历填充助手"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "rapidocr-onnxruntime>=1.3.0",
    "onnxruntime>=1.15.0",
    "PyYAML>=6.0",
    "python-Levenshtein>=0.21.0",
    "keyboard>=0.13.5",
    "pyperclip>=1.8.2",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "PyInstaller>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "mypy>=1.19.1",
    "ruff>=0.14.11",
]
